+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ Release 0.15.0 (unreleased)
 ==============================
 
 * Add ``replay-patches`` command to step through patch contributions interactively (#1290)
+* Add ``update-patch --patch`` to target an intermediate patch, rebasing later ones on top (#1335)
 
 Release 0.14.4 (released 2026-08-28)
 ====================================